@@ -1,9 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
+    compileSdkVersion 28
 
-    //useLibrary 'org.anothermonitor'
+    lintOptions {
+        checkReleaseBuilds false
+        // Or, if you prefer, you can continue to check for errors in release builds,
+        // but continue the build even when errors are found:
+        abortOnError false
+    }
 
     dexOptions {
         javaMaxHeapSize "4g"
@@ -11,11 +16,11 @@ android {
 
     defaultConfig {
         applicationId "org.owasp.seraphimdroid"
-        minSdkVersion 16 //Junfan should be 14 motify the another app
-        targetSdkVersion 22 //23
+        minSdkVersion 14 //Junfan should be 14 motify the another app
+        targetSdkVersion 26 //23
         compileOptions {
-            sourceCompatibility JavaVersion.VERSION_1_7
-            targetCompatibility JavaVersion.VERSION_1_7
+            sourceCompatibility JavaVersion.VERSION_1_8
+            targetCompatibility JavaVersion.VERSION_1_8
         }
 
         multiDexEnabled true
@@ -24,6 +29,7 @@ android {
     buildTypes {
         release {
             minifyEnabled false
+//            multiDexKeepFile file('multidex-config.txt')
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
     }
@@ -32,15 +38,16 @@ android {
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
 
-    api 'com.android.support:appcompat-v7:21+'
-    api 'com.android.support:cardview-v7:+'
-    api 'com.android.support:recyclerview-v7:21+'
-    api 'com.android.support:support-v4:23.0.0'
+    api 'androidx.appcompat:appcompat:1.0.0'
+
+    api 'androidx.cardview:cardview:1.0.0'
+    api 'androidx.recyclerview:recyclerview:1.0.0'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
 
     api 'com.google.android.gms:play-services:6.1+'
     api 'com.google.code.gson:gson:2.3.1'
 
-    api 'com.android.support:multidex:1.0.0'
+    api 'androidx.multidex:multidex:2.0.0'
 
     api 'com.android.volley:volley:1.0.0'
     api(project(':lucenedroid')) {
@@ -48,4 +55,6 @@ dependencies {
     }
 
     implementation project(path: ':AnotherMonitor', configuration:'default')
+    implementation project(path: ':networkmonitor')
+
 }
